rag_api/requirements.txt
flask
flask-cors
chromadb
sentence-transformers
google-generativeai
python-dotenv
numpy
pandas
markdown
pydantic
pymongo
bcrypt
flask-jwt-extended